feat(ci): use `rust-toolchain.toml`
