+++ .github/workflows/release-crates.yml
@@ -12,7 +12,6 @@ jobs:
       matrix:
         include:
           - os: ubuntu-20.04
-            rust-target: x86_64-unknown-linux-gnu
             platform: linux
             arch: x64
             regular_build: 'true'
@@ -20,37 +19,27 @@ jobs:
     runs-on: ${{ matrix.os }}
     env:
       target: ${{ matrix.platform }}-${{ matrix.arch }}
-      isRelease: ${{ (startsWith(github.ref, 'refs/tags/') && (!contains(github.ref, 'rc') && (endsWith(github.ref, '0') || endsWith(github.ref, '2') || endsWith(github.ref, '4') || endsWith(github.ref, '6') || endsWith(github.ref, '8')))) }}
-      isNightly: ${{ ((startsWith(github.ref, 'refs/tags/') && !((!contains(github.ref, 'rc') && (endsWith(github.ref, '0') || endsWith(github.ref, '2') || endsWith(github.ref, '4') || endsWith(github.ref, '6') || endsWith(github.ref, '8'))))) || (!startsWith(github.ref, 'refs/tags/') && matrix.regular_build == 'true')) }}
-      isTest: ${{ matrix.rust-target == 'x86_64-unknown-linux-gnu' || matrix.rust-target == 'x86_64-pc-windows-msvc' }}
-      isUniversal: ${{ matrix.rust-target == 'x86_64-unknown-linux-gnu' }}
       CARGO_REGISTRY_TOKEN: ${{ secrets.CRATES_IO_TOKEN }}
     steps:
       - name: "Print Env"
         run: |
           echo "Running on ${{ matrix.os }}"
           echo "Target: ${{ env.target }}"
-          echo "Is Release: ${{ fromJson(env.isRelease) }}"
-          echo "Is Nightly: ${{ fromJson(env.isNightly) }}"
-          echo "Is Test: ${{ fromJson(env.isTest) }}"
-          echo "Is Universal (No Server): ${{ fromJson(env.isUniversal) }}"
       - uses: actions/checkout@v4
         with:
           submodules: recursive
-        if: (fromJson(env.isRelease) || fromJson(env.isNightly))
-      - name: rust toolchain
-        uses: dtolnay/rust-toolchain@1.84.0
-        with:
-          targets: ${{ matrix.rust-target }}
-        if: (fromJson(env.isRelease) || fromJson(env.isNightly))
+      # https://github.com/dtolnay/rust-toolchain/issues/133
+      # Only needed if your action will run two or more rust
+      # commands concurrently, otherwise rustup will lazily 
+      # install your rust-toolchain.toml when needed:
+      - name: 'Install from rust-toolchain.toml'
+        run: rustup show
       - name: Install llvm
-        if: matrix.platform == 'linux' && (fromJson(env.isRelease) || fromJson(env.isNightly))
         run: |
           sudo apt-get update
           sudo apt-get install llvm
       # - name: Run rust-cache
       #   uses: Swatinem/rust-cache@v2
-      #   if: (fromJson(env.isRelease) || fromJson(env.isNightly))
       - name: Publish crates
         run: |
           cargo publish --no-verify -p typst-shim || true
